+++ packages/scss/test/feature/index.test.js
@@ -1,11 +1,11 @@
 //--------------------------------------------------------
-//-- nwayo - Feature tests
+//-- SCSS - Feature tests
 //--------------------------------------------------------
 import configTests           from '../../../../test/feature/config-tests';
 import { given, when, then } from './index.gwt';
 
 
-describe(`Validate that 'nwayo' works`, () => {
+describe(`Validate that 'scss' works`, () => {
 
 	configTests({ given, when, then });