+++ packages/scss/test/standards/index.gwt.js
@@ -1,8 +1,8 @@
 //--------------------------------------------------------
-//-- nwayo - Given-When-Then
+//-- SCSS - Given-When-Then
 //--------------------------------------------------------
 import path     from 'path';
-import * as gwt from '../../../../test/feature/config.gwt';
+import * as gwt from '../../../../test/standards/package.gwt';
 
 const given = { ...gwt.given };
 const when  = { ...gwt.when };
@@ -10,7 +10,7 @@ const then  = { ...gwt.then };
 
 
 given.currentRoot = () => {
-	given.root(path.join(__dirname, '..', '..', 'dist', 'node'));
+	given.root(path.join(__dirname, '..', '..'));
 };